% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdgazer.R
\name{rdgazer}
\alias{rdgazer}
\title{Function for putting rdrobust output into stargazer format}
\usage{
rdgazer(rdlist, dvlabs = c(1:length(rdlist)), xlines = NULL, ...)
}
\arguments{
\item{rdlist}{takes a list of rdrobust objects}
}
\value{
a stargazer object
}
\description{
Function for putting rdrobust output into stargazer format
}
