% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdquant.R
\name{rdquant}
\alias{rdquant}
\title{Quantile regression in regression discontinuity framework 
see Frandsen, Frolich and Melly (2010). "Quantile Treatment Effects in the Regression Discontinuity Design"}
\usage{
rdquant(Y, x, fuzzy = NULL, c = 0, ...)
}
\arguments{
\item{Y}{a numeric vector}

\item{x}{a numeric vector}

\item{fuzzy}{a binary vector indicating dummy for treatment}

\item{c}{a double for the cutoff point}
}
\value{
a data frame of y values, coefficients, and standard errors
}
\description{
Quantile regression in regression discontinuity framework 
see Frandsen, Frolich and Melly (2010). "Quantile Treatment Effects in the Regression Discontinuity Design"
}
