\name{rdrobust-internal}
\alias{rdrobust-internal}
\alias{bwconst}
\alias{kweight}
\alias{qr.XXinv}
\alias{qr.reg}
\alias{rdvce}
\alias{regconst}
\title{Package internal functions}
\description{
Internal functions which are not part of the package application programming interface. 
}
\author{
Sebastian Calonico, University of Miami, Coral Gables, FL.  \email{scalonico@bus.miami.edu}.

Matias D. Cattaneo, University of Michigan, Ann Arbor, MI.  \email{cattaneo@umich.edu}.

Rocio Titiunik, University of Michigan, Ann Arbor, MI.  \email{titiunik@umich.edu}.
}

