% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdquant.R
\name{rdquant}
\alias{rdquant}
\title{Quantile regression in regression discontinuity framework 
see Frandsen, Frolich and Melly (2010). "Quantile Treatment Effects in the Regression Discontinuity Design"}
\usage{
rdquant(Y, grid = 50, ...)
}
\arguments{
\item{Y}{a numeric vector}

\item{grid}{an integer determining how finely to estimate pdfs}
}
\value{
a data frame of y values, coefficients, and standard errors
}
\description{
Quantile regression in regression discontinuity framework 
see Frandsen, Frolich and Melly (2010). "Quantile Treatment Effects in the Regression Discontinuity Design"
}
