\name{rdrobust-internal}
\alias{rdrobust-internal}
\alias{rdrobust_kweight}
\alias{rdrobust_vce}
\alias{rdrobust_res}
\alias{rdrobust_bw}
\alias{qrXXinv}
\alias{qrreg}
\alias{bwconst}
\alias{rdvce}
\alias{regconst}
\title{Package Internal Functions}
\description{
Internal functions which are not part of the package application programming interface. 
}
\author{
Sebastian Calonico, University of Miami, Coral Gables, FL.  \email{scalonico@bus.miami.edu}.

Matias D. Cattaneo, University of Michigan, Ann Arbor, MI.  \email{cattaneo@umich.edu}.

Max H. Farrell, University of Chicago, Chicago, IL. \email{max.farrell@chicagobooth.edu}.

Rocio Titiunik, University of Michigan, Ann Arbor, MI.  \email{titiunik@umich.edu}.
}

