\name{rdrobust_RDsenate} 
\alias{rdrobust_RDsenate}
\docType{data}
\title{RD Senate Data}
\description{
Extract of the dataset constructed by Cattaneo, Frandsen, and Titiunik (2014), which include measures of incumbency advantage in the U.S. Senate for the period 1914-2010.
}
\usage{data(rdrobust_RDsenate)}
\format{
  A data frame with 1390 observations on the following 2 variables.
  \describe{
    \item{\code{margin}}{a numeric vector.}
    \item{\code{vote}}{a numeric vector.}
  }
}
\source{
    Cattaneo, M. D., Frandsen, B., and R. Titiunik. 2014. "Randomization Inference in the Regression Discontinuity Design: An Application to the Study of Party Advantages in the U.S. Senate." \emph{Journal of Causal Inference}, forthcoming.  \url{http://www-personal.umich.edu/~cattaneo/papers/Cattaneo-Frandsen-Titiunik_2014_JCI.pdf}.
}
\references{
    Cattaneo, M. D., Frandsen, B., and R. Titiunik. 2014. "Randomization Inference in the Regression Discontinuity Design: An Application to the Study of Party Advantages in the U.S. Senate." \emph{Journal of Causal Inference}, forthcoming.  \url{http://www-personal.umich.edu/~cattaneo/papers/Cattaneo-Frandsen-Titiunik_2014_JCI.pdf}.
}
\keyword{datasets}
